move test db setup to conftest.py
